Remove RelationAdapters etc
Remove RelationAdapters  and add contexts to relation handlers.
Other thing of note is that default handlers are registered
based on a charms metadata.
